+++ examples/nuxtjs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @livekit/component-example-nuxt
 
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @livekit/components-vue@0.1.5
+
 ## 0.1.4
 
 ### Patch Changes